refactor(tokens): cleanup configs
